added solutions for  Populating The Populations

+++ labs_for_js_postgres_client/knex_labs/db/migrations/20200618142859_CreatePopulations.js
@@ -3,7 +3,7 @@ exports.up = function (knex) {
     return knex.schema.createTable("populations", (table) => {
       table.increments("id");
       table.integer("year");
-      table.integer("quantity");
+      table.bigInteger("quantity");
       table.bigInteger("country_id").references("countries.id");
     });
   };
